# Speed up `JavaImpl.canonizeFinite`