@agdavydov81 Thanks for the great work of maintaning this library!

I would like to propose a series of performance improvements that will be borrowed from [jsoniter-scala](https://github.com/plokhotnyuk/jsoniter-scala).

The first one is using 32-bit arithmetic when `long` values fit `int`.

The proposed change in this PR speeds up `Decimal64Utils.canonize` in ~1.5x times on Intel CPUs.

Would you be interested to propagate it to other places like `JavaImpl.fromDecimalDouble`, `JavaImpl.dropTrailingZeros`?