@agdavydov81 2023-06-18
Hi @plokhotnyuk !
Sorry for weekend delay.
Thank you very mush for your contribution.
Your ~1.5 speedup looks very interesting.
IMHO the speed up factor will depend from the mantissa size, so could you please commit the benchmark you used.
And of course, I'm very interested in your enhancements.