@plokhotnyuk 2023-06-22
I've used you benchmark modified a bit:
```
package com.epam.deltix.dfp;

import org.openjdk.jmh.annotations.*;
import org.openjdk.jmh.runner.Runner;
import org.openjdk.jmh.runner.RunnerException;
import org.openjdk.jmh.runner.options.Options;
import org.openjdk.jmh.runner.options.OptionsBuilder;

import java.util.concurrent.TimeUnit;

@BenchmarkMode(Mode.Throughput)
@OutputTimeUnit(TimeUnit.SECONDS)
@Warmup(time = 1, iterations = 3)
@Measurement(time = 1, iterations = 3)
@State(Scope.Thread)
@Fork(3)
public class CanonizeBenchmark {
    /* 0., 1., 1000., 1000000., 1000000000. */
    @Param({"3584865303386914816", "3584865303386914817", "3584865303386915816", "3584865303387914816", "3584865304386914816"})
    private long decimalValue;

    @Benchmark
    public long canonize() {
        return Decimal64Utils.canonize(decimalValue);
    }

    public static void main(final String[] args) throws RunnerException {
        final Options opt = new OptionsBuilder()
            .include(".*" + CanonizeBenchmark.class.getSimpleName() + ".*")
            .forks(1)
            .build();
        new Runner(opt).run();
    }
}
```
Below are results from Intel Core i7 11700.
### JDK-17
#### Before
```
Benchmark                        (decimalValue)   Mode  Cnt           Score         Error  Units
CanonizeBenchmark.canonize  3584865303386914816  thrpt    9  1617737671.603 ± 4384045.033  ops/s
CanonizeBenchmark.canonize  3584865303386914817  thrpt    9   653369891.264 ± 1195424.387  ops/s
CanonizeBenchmark.canonize  3584865303386915816  thrpt    9   164467021.789 ±  451702.106  ops/s
CanonizeBenchmark.canonize  3584865303387914816  thrpt    9   100039684.866 ±  130271.557  ops/s
CanonizeBenchmark.canonize  3584865304386914816  thrpt    9    69282795.208 ±  124951.509  ops/s
```
#### After
```
Benchmark                        (decimalValue)   Mode  Cnt           Score         Error  Units
CanonizeBenchmark.canonize  3584865303386914816  thrpt    9  1620055317.839 ± 3216821.802  ops/s
CanonizeBenchmark.canonize  3584865303386914817  thrpt    9   670064646.442 ± 1226980.219  ops/s
CanonizeBenchmark.canonize  3584865303386915816  thrpt    9   195484212.871 ±  634727.166  ops/s
CanonizeBenchmark.canonize  3584865303387914816  thrpt    9   129336384.372 ±  134439.630  ops/s
CanonizeBenchmark.canonize  3584865304386914816  thrpt    9    95455222.498 ±  201239.784  ops/s
```
### Oracle GraalVM JDK-20
#### Before
```
Benchmark                        (decimalValue)   Mode  Cnt           Score          Error  Units
CanonizeBenchmark.canonize  3584865303386914816  thrpt    9  1616625045.114 ±  2716094.139  ops/s
CanonizeBenchmark.canonize  3584865303386914817  thrpt    9   841671652.079 ± 10986388.798  ops/s
CanonizeBenchmark.canonize  3584865303386915816  thrpt    9   292773953.152 ±   551339.210  ops/s
CanonizeBenchmark.canonize  3584865303387914816  thrpt    9   174697012.191 ±   265054.063  ops/s
CanonizeBenchmark.canonize  3584865304386914816  thrpt    9   117475748.825 ±   271808.566  ops/s
```
#### After
```
Benchmark                        (decimalValue)   Mode  Cnt           Score         Error  Units
CanonizeBenchmark.canonize  3584865303386914816  thrpt    9  1616472006.114 ± 2070571.039  ops/s
CanonizeBenchmark.canonize  3584865303386914817  thrpt    9  1057016669.674 ± 9394885.699  ops/s
CanonizeBenchmark.canonize  3584865303386915816  thrpt    9   404299358.429 ± 1040294.081  ops/s
CanonizeBenchmark.canonize  3584865303387914816  thrpt    9   263831506.745 ±  645382.270  ops/s
CanonizeBenchmark.canonize  3584865304386914816  thrpt    9   176650708.266 ±  589620.206  ops/s
```