@agdavydov81 2023-06-23
Hi!
I have commited your fixes to the div10-speedup branch.
I have made some changes to your code, please review.